ファイルサイズの上限変更

+++ app/Http/Controllers/RecipeController.php
@@ -163,8 +163,9 @@ class RecipeController extends Controller
             // アップロードされたファイル名を取得
             // $file_name = $request->file('recipe.image')->getClientOriginalName();
             // 元々登録されていた画像は削除
-            Storage::disk('sftp')->delete($recipe->image);
+            if($recipe->image != null) Storage::disk('sftp')->delete($recipe->image);
             $file = $request->file('recipe.image');
+            // dd(is_readable(Storage::disk('sftp')->putFile($userId, $file)));
             if(!\Storage::disk('sftp')->exists($userId)) {
                 \Storage::disk('sftp')->makeDirectory($userId);
             }
@@ -307,7 +308,7 @@ class RecipeController extends Controller
     public function recipe_delete(Recipe $recipe)
     {
         // 元々登録されていた画像は削除
-        Storage::disk('sftp')->delete($recipe->image);
+        if($recipe->image != null) Storage::disk('sftp')->delete($recipe->image);
         $recipe->delete();
         return redirect('/recipes');
     }
